packages/python/anip-grpc/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anip-grpc"
version = "0.11.0"
description = "ANIP gRPC transport binding"
requires-python = ">=3.11"
dependencies = [
    "anip-service==0.11.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/anip_grpc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
